+++ resources/views/bootstrap-components/media/audio.blade.php
@@ -2,7 +2,7 @@
     @include('bootstrap-components::bootstrap-components.partials.label', ['labelClasses' => ['d-block']])
     <audio{{ htmlAttributes($componentId ? ['id' => $componentId] : null) }}{{ classTag('component', $componentClasses) }}{{ htmlAttributes($componentHtmlAttributes) }}>
         <source{{ htmlAttributes($src ? 'src="'.$src.'"' : null) }}>
-        @lang('bootstrap-components::bootstrap-components.notification.audio')
+        @lang('Your browser does not support the :tag HTML5 tag.', ['tag' => 'audio'])
     </audio>
     @include('bootstrap-components::bootstrap-components.partials.legend')
 </div>