+++ resources/views/bootstrap-components/media/video.blade.php
@@ -2,7 +2,7 @@
     @include('bootstrap-components::bootstrap-components.partials.label', ['labelClasses' => ['d-block']])
     <video{{ htmlAttributes($componentId ? ['id' => $componentId] : null) }}{{ classTag('component', $componentClasses) }}{{ htmlAttributes($poster ? 'poster="'.$poster.'"' : null) }}{{ htmlAttributes($componentHtmlAttributes) }}>
         <source{{ htmlAttributes($src ? 'src="'.$src.'"' : null) }}>
-        @lang('bootstrap-components::bootstrap-components.notification.video')
+        @lang('Your browser does not support the :tag HTML5 tag.', ['tag' => 'video'])
     </video>
     @include('bootstrap-components::bootstrap-components.partials.legend')
 </div>