+++ src/Components/Buttons/Update.php
@@ -43,7 +43,7 @@ class Update extends SubmitAbstract
      */
     protected function setLabel(): ?string
     {
-        return (string) __('bootstrap-components::bootstrap-components.label.update');
+        return (string) __('Update');
     }
 
     /**