+++ tests/Unit/Buttons/Abstracts/ButtonTestAbstract.php
@@ -69,14 +69,6 @@ abstract class ButtonTestAbstract extends SubmitTestAbstract
         $this->assertStringNotContainsString('<span class="label">default-label</span>', $html);
     }
 
-    public function testSetTranslatedLabel()
-    {
-        $label = 'bootstrap-components::bootstrap-components.label.validate';
-        $html = $this->getComponent()->label($label)->toHtml();
-        $this->assertStringContainsString('title="' . __($label) . '">', $html);
-        $this->assertStringContainsString('<span class="label">' . __($label) . '</span>', $html);
-    }
-
     public function testNoLabel()
     {
         $html = $this->getComponent()->label(null)->toHtml();